refactor(opentrons-ai-client): refactor InputPrompt phase1
refactor InputPrompt phase1 export functions to resources/utils from InputPrompt

close AUTH-2173 partially
