+++ opentrons-ai-client/src/resources/hooks/useInputPromptController.ts
@@ -1,0 +1,17 @@
+type UseInputPromptControllerArgs = {
+  userPrompt: string
+  resetForm: () => void
+  setUserPrompt: (v: string) => void
+}
+
+
+type UseInputPromptControllerResult = {
+  submitChat: () => void
+  submitCreateOrUpdate: () => void
+  isLoading: boolean
+  error: unknown
+  attachedFiles: File[]
+  fileError: string | null
+  handleFileSelect: (files: FileList | null) => void
+  handleRemoveFile: (index: number) => void
+}
