+++ opentrons-ai-client/src/resources/utils/calcTextAreaHeight.ts
@@ -1,0 +1,9 @@
+/**
+ * Calculates the number of lines in a given string.
+ * @param input - The string to calculate the number of lines for.
+ * @returns The number of lines in the input string.
+ */
+
+export const calcTextAreaHeight = (input: string): number => {
+  return input.split('\n').length
+}
